test: check the STM32H7 clock tree against the DS12110 bus limits
Compile platform/mcu/STM32H7xx/drivers/rcc.cpp in the Linux unit test build
against a fake stm32h7xx.h whose RCC/PWR/FLASH register blocks live in RAM
(layouts, bit positions and base addresses copied from stm32h743xx.h; the
ready flags mirror their enable bits so the busy-waits in startPll()
terminate). The test runs startPll(), decodes SYSCLK/HCLK/PCLK1-4 from the
programmed PLL1 and prescaler fields the way the hardware does, and checks:

- every bus is within the DS12110 VOS1 limits (400/200/100MHz);
- getBusClock() agrees with the programmed prescalers for AHB and APB1-4,
  the invariant PR #330 broke, and rcc_getPeriphClock() maps peripheral
  addresses to the right bus;
- USART6 and SPI4/5 stay on pclk2, the GPS BRR comes out as 10417 at
  9600 Bd, the flash SPI4 divider still yields 25MHz and the APB2 timer
  clock (TIM8 backlight PWM) stays at 200MHz with TIMPRE = 0.

rcc_getPeriphClock() now casts the peripheral pointer through uintptr_t:
reinterpret_cast<uint32_t> of a 64-bit host pointer is ill-formed, and the
result is identical on the 32-bit target.

+++ platform/mcu/STM32H7xx/drivers/rcc.h
@@ -51,7 +51,7 @@ uint32_t getBusClock(const uint8_t bus);
  */
 static constexpr uint32_t rcc_getPeriphClock(const void *periph)
 {
-    uint32_t addr = reinterpret_cast<uint32_t>(periph);
+    uint32_t addr = reinterpret_cast<uintptr_t>(periph);
 
     switch(addr & 0xFFFF0000) {
         case APB1PERIPH_BASE:
